infratesting/tests/postplan/requirements.txt
attrs==22.2.0
exceptiongroup==1.1.0
iniconfig==2.0.0
packaging==23.0
pluggy==1.0.0
pyhcl==0.4.4
pytest==7.2.1
tomli==2.0.1
